ChatGPT/requirements.txt
openai==0.27.3
python-dotenv==0.19.2